requirements.txt
streamlit
PyPDF2
langchain
langchain-community
faiss-cpu
sentence-transformers
numpy
tqdm
python-dotenv
